+++ src/main/java/org/opentripplanner/standalone/config/BuildConfig.java
@@ -303,6 +303,11 @@ public class BuildConfig {
     public final StorageConfig storage;
 
     /**
+     * Visibility calculations for an area will not be done if there are more nodes than this limit.
+     */
+    public final int maxAreaNodes;
+
+    /**
      * Set all parameters from the given Jackson JSON tree, applying defaults.
      * Supplying MissingNode.getInstance() will cause all the defaults to be applied.
      * This could be done automatically with the "reflective query scraper" but it's less type safe and less clear.
@@ -349,6 +354,7 @@ public class BuildConfig {
         transitServiceEnd = c.asDateOrRelativePeriod( "transitServiceEnd", "P3Y");
         useTransfersTxt = c.asBoolean("useTransfersTxt", false);
         writeCachedElevations = c.asBoolean("writeCachedElevations", false);
+        maxAreaNodes = c.asInt("maxAreaNodes", 500);
 
         // List of complex parameters
         fareServiceFactory = DefaultFareServiceFactory.fromConfig(c.asRawNode("fares"));